backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
django-filter==23.3
djangorestframework-simplejwt==5.3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
werkzeug==3.0.1
pymysql==1.1.0
scikit-learn==1.3.2
mysql-connector-python==8.0.33
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
